+++ NEWS.md
@@ -3,8 +3,9 @@
 - *Package* 
   - `H2cal()`: BLUEs from H2cal may be wrong if other fixed effects in the model (#10).
     - Thanks Jim Holland (@ncsumaize) for the suggestion to improve the function.
-  - New function: `gdoc_rticle()`
-    - Use Articul8 Add-ons from Google docs to build a Rticles
+  - New function: `gdoc2rmd()`
+    - Use Articul8 Add-ons from Google docs to build Rticles
+  - Update pkgdown
 
 # Version 0.5.0